Make compatible with private maxmind database delivery
This introduces a number of changes aimed at making ip-geoloc work with
MaxMind's new individually licensed distribution model.

Since the URLs for downloading the database are no longer public:
* the `:database-md5-url` and `:database-url` config keys were removed.
* the `:maxmind-license-key` config key was added to enable building the
  private versions of the afformenteioned urls.

Maxmind has stopped providing MD5 fingerprints of downloads and now
offer them as a SHA256. Additionally, the format the fingerprint is
available in now includes the file name in the returned value, so the
method of extracting the fingerprint has changed slightly, as has the
naming.

The database download now comes as part of a tar'd gzip'd archive that
includes a directory of files. Since the download is now tar'd (where
before it was not), this necessitated the introduction of a
decompression lib that understood the tar format. I settled on
clj-compress. Additionally, since it's no longer possible to simply
decompress the database file and rename it, the algorithm now iterates
the tar entries and finds the mmdb file inside the decompressed
directory for use. This changes some aspects of the algorithm for
updating the database and some particulars of its storage on disk, but
nothing substantive beyond that.

Obvious changes were made to the README and code examples, but more are
likely appropriate.